🚀 5단계 - 자동차 경주(리팩터링)
