+++ README.md
@@ -19,3 +19,7 @@
 * [x] 전진하는 자동차를 출력할 때 자동차 이름을 같이 출력한다.
 * [x] 자동차 이름은 쉼표(,)를 기준으로 구분한다.
 * [x] 자동차 경주 게임을 완료한 후 누가 우승했는지를 알려준다. 우승자는 한 명 이상일 수 있다.
+
+## Step5 요구사항
+* [x] 핵심 비지니스 로직을 가지는 객체를 domain 패키지, UI 관련한 객체를 view 패키지에 구현한다.
+* [x] MVC 패턴 기반으로 리팩토링해 view 패키지의 객체가 domain 패키지 객체에 의존할 수 있지만, domain 패키지의 객체는 view 패키지 객체에 의존하지 않도록 구현한다.
