+++ src/main/kotlin/racingcar/RacingGameApplication.kt
@@ -1,0 +1,13 @@
+package racingcar
+
+import racingcar.controller.RacingGame
+import racingcar.view.InputView
+
+fun main() {
+    val carNames = InputView().inputCarNames()
+    val tryNum = InputView().inputTryNum()
+
+    val racingGame = RacingGame(carNames)
+    println("\n실행 결과")
+    racingGame.start(tryNum)
+}
