+++ src/main/kotlin/racingcar/domain/ConditionGenerator.kt
@@ -1,0 +1,10 @@
+package racingcar.domain
+
+import kotlin.random.Random
+import kotlin.random.nextInt
+
+interface ConditionGenerator {
+    fun generate(): Int {
+        return Random.nextInt(0..9)
+    }
+}
