+++ src/main/kotlin/racingcar/domain/RandomConditionGenerator.kt
@@ -1,0 +1,3 @@
+package racingcar.domain
+
+class RandomConditionGenerator : ConditionGenerator
