+++ src/main/kotlin/racingcar/view/InputView.kt
@@ -1,0 +1,13 @@
+package racingcar.view
+
+class InputView {
+    fun inputCarNames(): List<String> {
+        println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).")
+        return readln().split(",")
+    }
+
+    fun inputTryNum(): Int {
+        println("시도할 횟수는 몇 회 인가요?")
+        return readln().toInt()
+    }
+}
