+++ src/test/kotlin/racingcar/controller/RacingGameTest.kt
@@ -1,7 +1,8 @@
-package racingcar
+package racingcar.controller
 
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
+import racingcar.domain.ConditionGenerator
 
 class RacingGameTest : FunSpec({
     context("모든 차가 전진하면 항상 모든 차가 우승한다. 우승자는 한 명 이상일 수 있다.") {
