+++ src/test/kotlin/racingcar/domain/CarsTest.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
